Switch pull-request-target to base branch of pull request (#28921)
When pull request target workflow runs, we used build scripts from
main - to couple it with github workflows, but GitHub actually uses
base branch now for the workflows, so we should also use the target
branch for that.

This PR changes it and also extracts the code to do the checkout
to a separate composite action.

(cherry picked from commit 50f20037e8793913c6da2b296462a744dd02e23c)